requirements/docs.txt
# This file is @generated by PDM.
# Please do not edit it manually.

alabaster==0.7.13
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3; python_full_version <= "3.11.2"
autodoc-pydantic==2.1.0
babel==2.14.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6; sys_platform == "win32" or os_name == "nt" or platform_system == "Windows"
devtools==0.12.2
docutils==0.20.1
exceptiongroup==1.2.0; python_version < "3.11"
executing==2.0.1
fastapi==0.110.0
h11==0.14.0
idna==3.6
imagesize==1.4.1
importlib-metadata==7.0.2; python_full_version < "3.10.2"
jinja2==3.1.3
markupsafe==2.1.5
mmh3==4.1.0
packaging==24.0
pottery==3.0.0
pydantic==2.6.4
pydantic-core==2.16.3
pydantic-settings==2.2.1
pygments==2.17.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1; python_version < "3.9"
pyyaml==6.0.1
redis==4.6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sphinx==7.1.2
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
starlette==0.36.3
typing-extensions==4.10.0
ulid-py==1.1.0
urllib3==1.26.18
uvicorn==0.28.1
zipp==3.18.1; python_full_version < "3.10.2"
